for settinsgs

+++ products/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class Product(models.Model):
     name = models.CharField(max_length=100)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
     quantity = models.IntegerField()
     image = models.ImageField(upload_to='images',null=True,blank = True)
     descriptions = models.CharField(max_length=500)
